[REFACTOR] usersService split to friendService and usersService
- friendService is separated from usersService because friendService is
  a feature that is only used in usersGateway

(#14)
