+++ src/users/users.controller.ts
@@ -7,8 +7,6 @@ import {
   ValidationPipe,
   UseGuards,
   Put,
-  Post,
-  Delete,
   Query,
   NotFoundException,
 } from '@nestjs/common';
@@ -27,7 +25,6 @@ import {
   ApiParam,
   ApiBody,
   ApiBadRequestResponse,
-  ApiCreatedResponse,
 } from '@nestjs/swagger';
 import { UserEntity } from './entities/user.entity';
 import { getUser } from 'src/auth/decorator/get-user.decorator';
@@ -39,6 +36,8 @@ import { getUser } from 'src/auth/decorator/get-user.decorator';
 @ApiUnauthorizedResponse({ description: 'Invalid access token' })
 export class UsersController {
   constructor(private usersService: UsersService) {}
+
+  // NOTE: GET METHOD
 
   @Get('me')
   @ApiOperation({ summary: '내 정보 조회' })
@@ -61,36 +60,6 @@ export class UsersController {
     return { message: '사용 가능한 이름입니다.' };
   }
 
-  // @Get('friends')
-  // @ApiOperation({ summary: '나의 모든 친구 정보 조회' })
-  // @ApiOkResponse({
-  //   description: '나의 모든 친구 정보를 얻는다',
-  //   type: [UserInfoDto],
-  // })
-  // async getMyFriends(@getUser() user: UserEntity) {
-  //   return await this.usersService.getFriendList(user);
-  // }
-
-  // @Get('accept')
-  // @ApiOperation({ summary: '친구 요청 수락' })
-  // @ApiQuery({ name: 'id', description: '친구 요청을 수락할 유저의 id' })
-  // async acceptFriend(
-  //   @getUser() user: UserEntity,
-  //   @Query('id') friendId: string,
-  // ) {
-  //   await this.usersService.setFriendShipStatus(user, friendId, 'accept');
-  // }
-  //
-  // @Get('reject')
-  // @ApiOperation({ summary: '친구 요청 거절' })
-  // @ApiQuery({ name: 'id', description: '친구 요청을 거절할 유저의 id' })
-  // async rejectFriend(
-  //   @getUser() user: UserEntity,
-  //   @Query('id') friendId: string,
-  // ) {
-  //   await this.usersService.setFriendShipStatus(user, friendId, 'reject');
-  // }
-
   @Get(':id')
   @ApiOperation({ summary: '유저 정보 조회' })
   @ApiParam({ name: 'id', description: '조회할 유저 ID' })
@@ -106,28 +75,7 @@ export class UsersController {
     return await this.usersService.getAllUserInfo(user);
   }
 
-  @Delete('friend/:id')
-  @ApiOperation({ summary: '친구 삭제' })
-  async deleteFriend(
-    @getUser() user: UserEntity,
-    @Param('id') friendId: string,
-  ) {
-    await this.usersService.deleteFriend(user, friendId);
-  }
-
-  // @Post('friend')
-  // @ApiOperation({ summary: '친구 요청' })
-  // @ApiCreatedResponse({ description: '성공' })
-  // @ApiNotFoundResponse({ description: '존재하지 않는 유저입니다.' })
-  // @ApiBadRequestResponse({ description: '이미 친구요청을 보냈습니다.' })
-  // @ApiBadRequestResponse({
-  //   description: '자기 자신을 친구로 추가할 수 없습니다.',
-  // })
-  // async addFriend(@getUser() user: UserEntity, @Body('id') friendId: string) {
-  //   await this.usersService.addFriend(user, friendId);
-  //   return { message: '성공' };
-  // }
-
+  // NOTE: PUT METHOD
   @Put('me')
   @UsePipes(ValidationPipe)
   @ApiOperation({ summary: '유저 정보 수정' })
